Add a base skeleton/foundation
I want to have classes. First thought is to be able to parse the expression. I plan to iterate over them and check element types. If it's a number I want to put it into an array (stack) and if it's an operator, evaluate the top two numbers and push the result back into the array.