Remove parser, separate operator logic handling from the calculator and rename calculator to reverse-polish-notation-calculator.
The parser was an early choice that didn't make sense, it's not really parsing anything. The calculator class was getting too big though and it seemed logical to separate out the operator functions. This is an area that we might want to change the implementation of in the future (not use built in javascript math for example). The rename is just to make it obvious it's an RPN calculator.